Renamed folders, updated README, added setup.py, and adjusted structure for consistency
