+++ InfoFuzzyNetwork/__init__.py
@@ -1,0 +1,2 @@
+# IFN/__init__.py
+from .IFN import IFN
